requirements.txt
python-dotenv==0.19.1
pynput==1.7.4
pyperclip==1.8.2
openai==0.27.0
tkinter==8.6
os==0.1.4
subprocess.run==0.0.8
pystray==0.17.3
Pillow==8.3.2
threading==0.0.4